Refactor: results and grading-instructions use same DNDAnswerContainer
