+++ lerna.json
@@ -1,6 +1,6 @@
 {
   "npmClient": "npm",
   "packages": ["packages/*"],
-  "version": "23.2.3-beta.0",
+  "version": "23.2.3-beta.1",
   "exact": true
 }
