+++ packages/core/src/types/ExamAnswer.ts
@@ -1,5 +1,7 @@
 export type QuestionId = number
 
+export type ExamPage = 'exam' | 'results' | 'grading-instructions'
+
 interface AnswerCommon {
   questionId: QuestionId
   value: string
