+++ packages/generator/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@digabi/exam-engine-generator",
-  "version": "23.2.3-beta.0",
+  "version": "23.2.3-beta.1",
   "main": "dist/index.js",
   "author": "Matriculation Examination Board, Finland",
   "license": "EUPL-1.1",
@@ -8,7 +8,7 @@
     "dist"
   ],
   "dependencies": {
-    "@digabi/exam-engine-core": "23.2.3-beta.0",
+    "@digabi/exam-engine-core": "23.2.3-beta.1",
     "libxmljs2": "^0.35.0",
     "lodash": "^4.17.15"
   },
