+++ packages/mastering/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@digabi/exam-engine-mastering",
-  "version": "23.2.3-beta.0",
+  "version": "23.2.3-beta.1",
   "main": "dist/index.js",
   "author": "Matriculation Examination Board, Finland",
   "license": "EUPL-1.1",
@@ -9,8 +9,8 @@
     "schema"
   ],
   "dependencies": {
-    "@digabi/exam-engine-core": "23.2.3-beta.0",
-    "@digabi/exam-engine-rendering": "23.2.3-beta.0",
+    "@digabi/exam-engine-core": "23.2.3-beta.1",
+    "@digabi/exam-engine-rendering": "23.2.3-beta.1",
     "@ffprobe-installer/ffprobe": "^2.0.0",
     "cloneable-readable": "^3.0.0",
     "compare-versions": "^6.0.0",
