+++ packages/rendering/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@digabi/exam-engine-rendering",
-  "version": "23.2.3-beta.0",
+  "version": "23.2.3-beta.1",
   "main": "./dist/index.js",
   "license": "EUPL-1.1",
   "files": [
@@ -13,7 +13,7 @@
     "@babel/polyfill": "^7.8.7",
     "@babel/preset-env": "^7.8.7",
     "@babel/runtime": "^7.8.7",
-    "@digabi/exam-engine-core": "23.2.3-beta.0",
+    "@digabi/exam-engine-core": "23.2.3-beta.1",
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
     "babel-loader": "^9.1.2",
     "child-process-promise": "^2.2.1",
